+++ activerecord/CHANGELOG.md
@@ -1284,4 +1284,24 @@
 
     *Yves Senn*
 
+*   Schema Migrations now contain time of creation information to track when
+    migrations were run.
+    ```
+	  database: vagrant_development
+
+	  Status   Migration ID    Migration Run             Migration Name
+	  ---------------------------------------------------------------------------
+	  up       20150918171109  2015-09-18 17:20:22 UTC   ********** NO FILE **********
+	  up       20150918210543  2015-09-18 21:07:10 UTC   Bar
+	  down     20150918210806                            Baz
+    ```
+
+*   In the environmnet file, you can now request that schema migrations be ordered by invocatio
+    time:
+
+      config.active_record.schema_migrations_by_invocation_time = true
+
+    When you use `rake db:migrate:status` the output will be ordered by the Migration Run column
+    When you use `rake db:rollback` migrations will be rolled back in the order they were inovoked!
+
 Please check [4-2-stable](https://github.com/rails/rails/blob/4-2-stable/activerecord/CHANGELOG.md) for previous changes.